api/requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Deprecated==1.2.12
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.65.1
fastapi-users[mongodb]
h11==0.12.0
httptools==0.1.2
idna==3.1
makefun==1.11.3
motor==2.4.0
passlib==1.7.4
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.11.4
python-dotenv==0.17.1
python-multipart==0.0.5
PyYAML==5.4.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
wrapt==1.12.1
motor
psycopg2
requests
pytest